BackendFlask/requirements.txt
Flask==3.1.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
flask-jwt-extended==4.6.0
marshmallow==3.22.0
PyMySQL==1.1.1
flask-jwt-extended==4.6.0
PyJWT==2.8.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.3.3